Created Code that reads US and IR sensors
I made the code incorporated both Sensors.

However, i'm not sure if the readings are sent to the UART properly. Also, i'm not 100% sure if I did the thermal sensor properly either.

This code may need to incorporate the ADC, but im not 100% sure.

main_git is our 1st code, main.ino and all other c and h files is the real code